refactor: split persistence facade auto-wiring under the module-size budget
